requirements.txt
python-docx==0.8.11
beautifulsoup4==4.12.2
ensure==1.0.2
mkdocs-material
-e .